--- src/main/java/com/fnp/checkexception/custom-checked-exception/CheckedExceptions.java
@@ -1,18 +1,0 @@
-import customexceptions.ObservedException;
-import customexceptions.ConditionWhichThrowsException;
-
-class CheckedExceptions{	
-	
-	public static void main(String [] args){
-		
-		try{
-			ConditionWhichThrowsException condition = new ConditionWhichThrowsException();
-			condition.conditionChecker(6);
-		}
-		catch(ObservedException e){
-			System.out.println("Observed Exception is caught .");
-			e.printStackTrace();
-		}
-		System.out.println("Control Moves Ahead");
-	}
-}
