+++ src/main/java/com/fnp/checkexception/custom-checked-exception/resourcecontainer/ObservedException.java
@@ -1,4 +1,4 @@
-package customexceptions;
+package resourcecontainer;
 
 public class ObservedException extends Exception{
 	public ObservedException(String info){
