+++ src/main/java/com/fnp/checkexception/unchecked-exceptions-program/resourcecontainer/Equality.java
@@ -1,0 +1,12 @@
+package resourcecontainer;
+
+public class NumberEquality{
+	  int num1 = 5;
+		public void numberChecker(int num2){  
+		if(num1 != num2){
+			throw new ObservedException("The condition does not follow. ");
+	
+		}
+				
+	}
+}
