+++ src/main/java/com/fnp/checkexception/unchecked-exceptions-program/resourcecontainer/ObservedException.java
@@ -1,4 +1,4 @@
-package customexception;
+package resourcecontainer;
 
 public class ObservedException extends RuntimeException{
 	public ObservedException(String info){
